Refactor authorsToString in a util object, and test it
